backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-genai
python-dotenv
